slettet html ubrugt

+++ src/main/java/com/example/smartslate/repository/IEmployeeTask.java
@@ -4,6 +4,7 @@ package com.example.smartslate.repository;
 import com.example.smartslate.model.EmployeeTask;
 import com.example.smartslate.model.Task;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface IEmployeeTask {
@@ -11,5 +12,5 @@ public interface IEmployeeTask {
 
     List<Task> getEmployeeTasksByUserId(int employeeID);
     String calculateTotalTimeSpent(List<Task> tasks);
-    String formatTotalTime(double totalTime);
+    String formatTotalTime(BigDecimal totalTime);
 }
